EffectClient
